V0.3.0

+++ docker/build.sh
@@ -2,30 +2,43 @@
 BUIILD_CONTAINER='faissdb-build'
 BASE_IMAGE='faissdb:base'
 BUILD_IMAGE='faissdb:build'
-docker rm -f "${BUIILD_CONTAINER}"
-docker rmi "${BASE_IMAGE}"
-docker rmi "${BUILD_IMAGE}"
+
+RETRY=false
+KEEP=false
+for arg in "$@"; do
+  case "$arg" in
+    retry) RETRY=true ;;
+    keep)  KEEP=true ;;
+  esac
+done
 
 set -e
 cd `dirname $0`
 
-rm -rf build/mnt
-cp -r mnt build/mnt
-# cp mnt/* build/mnt/
-# cp mnt/.faissdb build/mnt/.faissdb
+if [ "$RETRY" != "true" ]; then
+  docker rm -f "${BUIILD_CONTAINER}"
+  docker rmi -f "${BASE_IMAGE}"
+  docker rmi -f "${BUILD_IMAGE}"
+  rm -rf build/mnt
+  cp -r mnt build/mnt
 
-cd build
-docker image build -t "${BASE_IMAGE}" .
-docker run --name="${BUIILD_CONTAINER}" -ti --tmpfs /run --tmpfs /run/lock --tmpfs /tmp:exec \
- -v /lib/modules:/lib/modules:ro \
- -v `pwd`/mnt:/mnt \
- -v `pwd`/../..:/mnt/faissdb \
- -d "${BASE_IMAGE}" \
- /bin/bash
+  cd build
+  docker image build -t "${BASE_IMAGE}" .
+  docker run --name="${BUIILD_CONTAINER}" -ti --tmpfs /run --tmpfs /run/lock --tmpfs /tmp:exec \
+   -v /lib/modules:/lib/modules:ro \
+   -v `pwd`/mnt:/mnt \
+   -v `pwd`/../..:/mnt/faissdb \
+   -v `pwd`/../../../go-faiss:/mnt/go-faiss \
+   -d "${BASE_IMAGE}" \
+   /bin/bash
+else
+  cp mnt/build.sh build/mnt/build.sh
+  cd build
+fi
 
 docker exec "${BUIILD_CONTAINER}" bash --login /mnt/build.sh
 
 docker commit "${BUIILD_CONTAINER}" "${BUILD_IMAGE}"
-if [ "$1" != "keep" ]; then
+if [ "$KEEP" != "true" ]; then
   docker rm -f "${BUIILD_CONTAINER}"
 fi
